pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mlx-websockets"
dynamic = ["version"]
description = "WebSocket streaming server for MLX models"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Lucas Johnston Kurilov", email = "code@lucasco.de"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: MacOS :: MacOS X",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mlx>=0.15.0",
    "mlx-lm>=0.15.0",
    "mlx-vlm>=0.0.6",
    "websockets>=12.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
]

[project.scripts]
mlx = "mlx_websockets.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "commitizen>=3.0.0",
]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
(
  venv/
  | __pycache__/
  | \.git/
  | build/
  | dist/
)
'''

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811", "E402"]  # Allow unused imports, redefinitions, and module-level imports not at top in tests
"mlx_websockets/server.py" = ["F401", "F841"]  # Allow unused imports and variables for optional dependencies

[tool.setuptools_scm]
write_to = "mlx_websockets/_version.py"
version_scheme = "post-release"
local_scheme = "no-local-version"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Timeout for tests to prevent hanging
timeout = 30
timeout_method = "thread"

# Markers
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "benchmark: marks tests as performance benchmarks"
]

# Logging
log_cli = true
log_cli_level = "WARNING"

filterwarnings = [
    "ignore::UserWarning:torchvision.datapoints",
    "ignore::UserWarning:torchvision.transforms.v2",
    "ignore::pytest.PytestDeprecationWarning:pytest_asyncio"
]
addopts = "--strict-markers -vv --cov=mlx_websockets --cov-report=html --cov-report=term-missing"
